udpcast-20060325
v20060325
       * Fix packaging, so that we now can build an clean rpm with
       just the command-line programs. The mkimage script for the
       embedded system is now within another RPM (udpcast-mkimage)
       which, for the moment, is still constructed the old kludgy
       way.
v20060320
       * Fixed signed-ness issues (revealed by gcc4)
       * By default write files in unsynchronized mode, for better performance
       * Fix case where data address is the same as rendez-vous address.

+++ socklib.c
@@ -311,7 +311,7 @@ int doReceive(int s, void *message, size_t len,
 
 int getSendBuf(int sock) {
     int bufsize;
-    int len=sizeof(int);
+    size_t len=sizeof(int);
     if(getsockopt(sock, SOL_SOCKET, SO_SNDBUF, (char*)&bufsize, &len) < 0)
 	return -1;
     return bufsize;
@@ -324,7 +324,7 @@ void setSendBuf(int sock, unsigned int bufsize) {
 
 unsigned int getRcvBuf(int sock) {
     unsigned int bufsize;
-    int len=sizeof(int);
+    size_t len=sizeof(int);
     if(getsockopt(sock, SOL_SOCKET, SO_RCVBUF, (char *)&bufsize, &len) < 0)
 	return -1;
     return bufsize;
